backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.6
tenacity==9.0.0
azure-search-documents==11.6.0
azure-core==1.32.0
pytest==8.3.4
pytest-asyncio==0.25.3